+++ spec/spec_helper.rb
@@ -16,7 +16,8 @@ RSpec.configure do |config|
   config.treat_symbols_as_metadata_keys_with_true_values = true
   config.run_all_when_everything_filtered = true
   config.filter_run :focus
-
-  # Run specs in random order to surface order dependencies.
-  config.order = 'random'
+  config.order = 'random' # Run specs in random order to surface order dependencies.
+  config.expect_with :rspec do |c|
+    c.syntax = :expect  # disable `should` syntax  http://goo.gl/BGxqP
+  end
 end